Refactor polygon decoder function
